+++ tests/upgrade_contract_test.php
@@ -54,6 +54,7 @@ final class upgrade_contract_test extends advanced_testcase {
         $this->assertStringContainsString('if ($oldversion < 2026060453)', $source);
         $this->assertStringContainsString('if ($oldversion < 2026082104)', $source);
         $this->assertStringContainsString("'serverplaybacksessionid'", $source);
+        $this->assertStringContainsString('focus_policy::ensure_exception_group', $source);
     }
 
     /**